+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [2.3.0]
+## Unreleased
 
 ### Added
+
+- Example to demonstrate how to create vertical BarSeries
 - Colorblind friendly Okabe-Ito color palette
+
+### Fixed
+- DateTimeAxis converting local time rather than UTC to TimeZone
 
 ## [2.2.0] - 2024-09-03